Break out for testability
